+++ src/main/kotlin/com/example/custard/domain/post/dto/info/DateInfo.kt
@@ -1,16 +1,16 @@
-package com.example.custard.domain.post.dto.response
+package com.example.custard.domain.post.dto.info
 
 import com.example.custard.domain.post.model.date.Date
 import java.time.LocalDate
 import java.time.LocalTime
 
-class PostDateResponse (
+class DateInfo(
     val date: LocalDate,
     val time: LocalTime?,
 ) {
     companion object {
-        fun of(date: Date): PostDateResponse {
-            return PostDateResponse(date.date, date.time)
+        fun toEntity(info: DateInfo): Date {
+            return Date(info.date, info.time)
         }
     }
 }
