+++ src/main/kotlin/com/example/custard/domain/post/dto/info/PostCreateInfo.kt
@@ -10,7 +10,7 @@ class PostCreateInfo (
     val categoryId: Long,
     val title: String,
     val description: String,
-    val dates: List<PostDateInfo>,
+    val dates: List<DateInfo>,
     val delivery: Boolean,
     val place: String?,
     val minPrice: Int,
@@ -21,7 +21,7 @@ class PostCreateInfo (
     constructor(categoryId: Long,
                 title: String,
                 description: String,
-                dates: List<PostDateInfo>,
+                dates: List<DateInfo>,
                 delivery: Boolean,
                 place: String?,
                 minPrice: Int,
@@ -31,7 +31,7 @@ class PostCreateInfo (
     constructor(categoryId: Long,
                 title: String,
                 description: String,
-                dates: List<PostDateInfo>,
+                dates: List<DateInfo>,
                 delivery: Boolean,
                 place: String?,
                 minPrice: Int,
