+++ src/main/kotlin/com/example/custard/domain/post/dto/request/PostCreateRequest.kt
@@ -1,7 +1,7 @@
 package com.example.custard.domain.post.dto.request
 
 import com.example.custard.domain.post.dto.info.PostCreateInfo
-import com.example.custard.domain.post.dto.info.PostDateInfo
+import com.example.custard.domain.post.dto.info.DateInfo
 import com.example.custard.domain.post.model.PostType
 
 class PostCreateRequest (
@@ -9,7 +9,7 @@ class PostCreateRequest (
     val categoryId: Long,
     val title: String,
     val description: String,
-    val dates: List<PostDateRequest>,
+    val dates: List<DateRequest>,
     val delivery: Boolean,
     val place: String?,
     val minPrice: Int,
@@ -17,7 +17,7 @@ class PostCreateRequest (
     val productId: Long?
 ) {
     fun createInfo(): PostCreateInfo {
-        val datesInfo: List<PostDateInfo> = dates.map { it.createInfo() }
+        val datesInfo: List<DateInfo> = dates.map { it.createInfo() }
         return when (type) {
             PostType.PURCHASE -> PostCreateInfo(categoryId, title, description, datesInfo, delivery, place, minPrice, maxPrice)
             PostType.SALE -> PostCreateInfo(categoryId, title, description, datesInfo, delivery, place, minPrice, maxPrice, productId)
