+++ src/main/kotlin/com/example/custard/domain/post/dto/response/PostDetailResponse.kt
@@ -10,7 +10,7 @@ class PostDetailResponse (
     val type: PostType,
     val title: String,
     val description: String,
-    val dates: List<PostDateResponse>,
+    val dates: List<DateResponse>,
     val delivery: Boolean,
     val place: String?,
     val minPrice: Int,
@@ -27,7 +27,7 @@ class PostDetailResponse (
                 type = post.type,
                 title = post.title,
                 description = post.description,
-                dates = post.dates.map { PostDateResponse.of(it.date) },
+                dates = post.dates.map { DateResponse.of(it.date) },
                 delivery = post.delivery,
                 place = post.place,
                 minPrice = post.minPrice,
